Round of enhancements to make APIs orthogonal.
- IntNumber will interpret json numbers as int64 value.
- Fixtojson to normalize language native transformations
  to json standard.
- value2cbor() will support json.Number (JsonNumber option).
- pass configurations to cbor2value().
- gson2collate() will support json.Number (JsonNumber option).

+++ value_json.go
@@ -122,7 +122,7 @@ func jsonnum2value(txt string, nk NumberKind) (string, interface{}) {
 		if err != nil {
 			panic("gson scanner expectedJsonInteger")
 		}
-		return txt[e:], num
+		return txt[e:], int64(num)
 	}
 	// FloatNumber, or FloatNumber32, or SmartNumber, or SmartNumber32
 	// NOTE: ignore the error because we have only picked
